+++ packages/extension-redis/src/Redis.ts
@@ -1,9 +1,8 @@
-import RedisClient, { ClusterNode, ClusterOptions, RedisOptions } from 'ioredis'
+import type { ClusterNode, ClusterOptions, RedisOptions } from 'ioredis'
+import RedisClient from 'ioredis'
 import Redlock from 'redlock'
 import { v4 as uuid } from 'uuid'
-import {
-  IncomingMessage,
-  OutgoingMessage,
+import type {
   Document,
   Extension,
   afterLoadDocumentPayload,
@@ -12,10 +11,12 @@ import {
   onStoreDocumentPayload,
   onAwarenessUpdatePayload,
   onChangePayload,
-  MessageReceiver,
-  Debugger,
   onConfigurePayload,
-  beforeBroadcastStatelessPayload, Hocuspocus,
+  beforeBroadcastStatelessPayload, Hocuspocus} from '@hocuspocus/server'
+import {
+  IncomingMessage,
+  OutgoingMessage,
+  MessageReceiver,
 } from '@hocuspocus/server'
 
 export type RedisInstance = RedisClient.Cluster | RedisClient.Redis
@@ -96,8 +97,6 @@ export class Redis implements Extension {
 
   locks = new Map<string, Redlock.Lock>()
 
-  logger: Debugger
-
   messagePrefix: Buffer
 
   /**
@@ -113,9 +112,6 @@ export class Redis implements Extension {
       ...this.configuration,
       ...configuration,
     }
-
-    // We’ll replace that in the onConfigure hook with the global instance.
-    this.logger = new Debugger()
 
     // Create Redis instance
     const {
@@ -151,7 +147,6 @@ export class Redis implements Extension {
   }
 
   async onConfigure({ instance }: onConfigurePayload) {
-    this.logger = instance.debugger
     this.instance = instance
   }
 
@@ -327,14 +322,11 @@ export class Redis implements Extension {
     const document = this.instance.documents.get(documentName)
 
     if (!document) {
-      // What does this mean? Why are we subscribed to this document?
-      this.logger.log(`Received message for unknown document ${documentName}`)
       return
     }
 
     new MessageReceiver(
       message,
-      this.logger,
       this.redisTransactionOrigin,
     ).apply(document, undefined, reply => {
       return this.pub.publishBuffer(