+++ packages/provider/src/OutgoingMessage.ts
@@ -1,5 +1,6 @@
-import { Encoder, createEncoder, toUint8Array } from 'lib0/encoding'
-import { MessageType, OutgoingMessageArguments, OutgoingMessageInterface } from './types.js'
+import type { Encoder} from 'lib0/encoding'
+import { createEncoder, toUint8Array } from 'lib0/encoding'
+import type { MessageType, OutgoingMessageArguments, OutgoingMessageInterface } from './types.js'
 
 export class OutgoingMessage implements OutgoingMessageInterface {
   encoder: Encoder