+++ packages/server/src/Connection.ts
@@ -1,15 +1,14 @@
-import { IncomingMessage as HTTPIncomingMessage } from 'http'
-import AsyncLock from 'async-lock'
-import WebSocket from 'ws'
+import type { IncomingMessage as HTTPIncomingMessage } from 'http'
+import type WebSocket from 'ws'
 import {
-  CloseEvent, Forbidden, WsReadyStates,
+  type CloseEvent, ResetConnection,
 } from '@hocuspocus/common'
-import Document from './Document.js'
+import { WsReadyStates} from '@hocuspocus/common'
+import type Document from './Document.js'
 import { IncomingMessage } from './IncomingMessage.js'
 import { OutgoingMessage } from './OutgoingMessage.js'
 import { MessageReceiver } from './MessageReceiver.js'
-import { Debugger } from './Debugger.js'
-import { onStatelessPayload } from './types.js'
+import type { onStatelessPayload } from './types.js'
 
 export class Connection {
 
@@ -29,11 +28,7 @@ export class Connection {
 
   socketId: string
 
-  lock: AsyncLock
-
-  readOnly: Boolean
-
-  logger: Debugger
+  readOnly: boolean
 
   /**
    * Constructor.
@@ -45,7 +40,6 @@ export class Connection {
     socketId: string,
     context: any,
     readOnly = false,
-    logger: Debugger,
   ) {
     this.webSocket = connection
     this.context = context
@@ -53,9 +47,6 @@ export class Connection {
     this.request = request
     this.socketId = socketId
     this.readOnly = readOnly
-    this.logger = logger
-
-    this.lock = new AsyncLock()
 
     this.webSocket.binaryType = 'nodebuffer'
     this.document.addConnection(this)
@@ -118,12 +109,6 @@ export class Connection {
     const message = new OutgoingMessage(this.document.name)
       .writeStateless(payload)
 
-    this.logger.log({
-      direction: 'out',
-      type: message.type,
-      category: message.category,
-    })
-
     this.send(
       message.toUint8Array(),
     )
@@ -133,14 +118,14 @@ export class Connection {
    * Graceful wrapper around the WebSocket close method.
    */
   close(event?: CloseEvent): void {
-    this.lock.acquire('close', (done: Function) => {
       if (this.document.hasConnection(this)) {
         this.document.removeConnection(this)
         this.callbacks.onClose.forEach((callback: (arg0: Document, arg1?: CloseEvent) => any) => callback(this.document, event))
-      }
 
-      done()
-    })
+        const closeMessage = new OutgoingMessage(this.document.name)
+        closeMessage.writeCloseMessage(event?.reason ?? 'Server closed the connection')
+        this.send(closeMessage.toUint8Array())
+      }
   }
 
   /**
@@ -154,12 +139,6 @@ export class Connection {
 
     const awarenessMessage = new OutgoingMessage(this.document.name)
       .createAwarenessUpdateMessage(this.document.awareness)
-
-    this.logger.log({
-      direction: 'out',
-      type: awarenessMessage.type,
-      category: awarenessMessage.category,
-    })
 
     this.send(awarenessMessage.toUint8Array())
   }
@@ -180,14 +159,13 @@ export class Connection {
       .then(() => {
         new MessageReceiver(
           message,
-          this.logger,
         ).apply(this.document, this)
       })
       .catch((e: any) => {
-        console.log('closing connection because of exception', e)
+        console.error('closing connection because of exception', e)
         this.close({
-          code: 'code' in e ? e.code : Forbidden.code,
-          reason: 'reason' in e ? e.reason : Forbidden.reason,
+          code: 'code' in e ? e.code : ResetConnection.code,
+          reason: 'reason' in e ? e.reason : ResetConnection.reason,
         })
       })
   }