+++ packages/server/src/util/getParameters.ts
@@ -1,4 +1,4 @@
-import { IncomingMessage } from 'http'
+import type { IncomingMessage } from 'http'
 import { URLSearchParams } from 'url'
 
 /**