+++ playground/frontend/src/pages/Index.vue
@@ -64,6 +64,7 @@ import Collaboration from '@tiptap/extension-collaboration'
 import {
   HocuspocusProvider, HocuspocusProviderWebsocket,
 } from '@hocuspocus/provider'
+import StatusBar from '../components/StatusBar.vue'
 
 const socket = new HocuspocusProviderWebsocket({
   url: 'ws://127.0.0.1:1234',
@@ -75,22 +76,18 @@ const $states2 = ref({})
 const provider = new HocuspocusProvider({
   websocketProvider: socket,
   name: 'hocuspocus-demo',
-  broadcast: true,
   onAwarenessUpdate: ({ states }) => {
     $states1.value = states
   },
 })
-provider.subscribeToBroadcastChannel()
 
 const anotherProvider = new HocuspocusProvider({
   websocketProvider: socket,
   name: 'hocuspocus-demo2',
-  broadcast: true,
   onAwarenessUpdate: ({ states }) => {
     $states2.value = states
   },
 })
-anotherProvider.subscribeToBroadcastChannel()
 
 const editor = new Editor({
   extensions: [
@@ -118,7 +115,7 @@ const anotherEditor = new Editor({
 
 </script>
 
-<style >
+<style>
 .ProseMirror {
   border: 1px solid grey;
   padding: 1rem;