+++ tests/utils/newHocuspocusProvider.ts
@@ -1,8 +1,10 @@
 import {
   HocuspocusProvider,
-  HocuspocusProviderConfiguration, HocuspocusProviderWebsocket, HocuspocusProviderWebsocketConfiguration,
+  type HocuspocusProviderConfiguration,
+  type HocuspocusProviderWebsocket,
+  type HocuspocusProviderWebsocketConfiguration,
 } from '@hocuspocus/provider'
-import { Hocuspocus } from '@hocuspocus/server'
+import type { Hocuspocus } from '@hocuspocus/server'
 import { newHocuspocusProviderWebsocket } from './newHocuspocusProviderWebsocket.js'
 
 export const newHocuspocusProvider = (
@@ -11,15 +13,14 @@ export const newHocuspocusProvider = (
   websocketOptions: Partial<HocuspocusProviderWebsocketConfiguration> = {},
   websocketProvider?: HocuspocusProviderWebsocket,
 ): HocuspocusProvider => {
-  return new HocuspocusProvider({
+  const provider = new HocuspocusProvider({
     websocketProvider: websocketProvider ?? newHocuspocusProviderWebsocket(server, websocketOptions),
     // Just use a generic document name for all tests.
     name: 'hocuspocus-test',
-    // There is no need to share data with other browser tabs in the testing environment.
-    broadcast: false,
-    // We don’t need console logging in tests. If we actually do, we can overwrite it anyway.
-    quiet: true,
     // Add or overwrite settings, depending on the test case.
     ...options,
   })
+  provider.attach()
+
+  return provider
 }